Compile all with NDK 26.3
With this version also minSdkVersion bring to 24 (Android 7).
New compiler detect the use of some functions defined in stdio.h
as fseeko available only from Android API Level >= 24 if __USE_FILE_OFFSET64
is enabled as in voip module.
The solution as described here
https://developer.android.com/ndk/guides/common-problems
is set minSdkVersion almost to 24 or remove __USE_FILE_OFFSET64